Pose Graph interface settled
